refactor: use `NonZeroU*` where appropriate (#208)
In particular, replace `u16` and `u32` for symbol counts, symbol sizes,
and number of shards by `NonZeroU16`.

+++ crates/walrus-core/benches/constants/mod.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-pub const N_SHARDS: u32 = 1000;
+pub const N_SHARDS: u16 = 1000;
 // Likely values for the number of source symbols for the primary and secondary encoding.
 // These values are consistent with BFT and are supported source-block sizes that do not require
 // padding, see https://datatracker.ietf.org/doc/html/rfc6330#section-5.6.
